moved common parts of db ActiveRelation and Query to traits
these parts can be reused in other ActiveRecord implementations
