+++ framework/yii/ar/ActiveQuery.php
@@ -1,0 +1,234 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\ar;
+
+use yii\db\ActiveRecord;
+
+/**
+ * ActiveQuery represents a DB query associated with an Active Record class.
+ *
+ * ActiveQuery instances are usually created by [[ActiveRecord::find()]], [[ActiveRecord::findBySql()]]
+ * and [[ActiveRecord::count()]].
+ *
+ * ActiveQuery mainly provides the following methods to retrieve the query results:
+ *
+ * - [[one()]]: returns a single record populated with the first row of data.
+ * - [[all()]]: returns all records based on the query results.
+ * - [[count()]]: returns the number of records.
+ * - [[scalar()]]: returns the value of the first column in the first row of the query result.
+ * - [[exists()]]: returns a value indicating whether the query result has data or not.
+ *
+ * Because ActiveQuery extends from [[Query]], one can use query methods, such as [[where()]],
+ * [[orderBy()]] to customize the query options.
+ *
+ * ActiveQuery also provides the following additional query options:
+ *
+ * - [[with()]]: list of relations that this query should be performed with.
+ * - [[indexBy()]]: the name of the column by which the query result should be indexed.
+ * - [[asArray()]]: whether to return each record as an array.
+ *
+ * These options can be configured using methods of the same name. For example:
+ *
+ * ~~~
+ * $customers = Customer::find()->with('orders')->asArray()->all();
+ * ~~~
+ *
+ * @author Qiang Xue <qiang.xue@gmail.com>
+ * @author Carsten Brandt <mail@cebe.cc>
+ * @since 2.0
+ */
+trait ActiveQuery
+{
+	/**
+	 * @var string the name of the ActiveRecord class.
+	 */
+	public $modelClass;
+	/**
+	 * @var array list of relations that this query should be performed with
+	 */
+	public $with;
+	/**
+	 * @var boolean whether to return each record as an array. If false (default), an object
+	 * of [[modelClass]] will be created to represent each record.
+	 */
+	public $asArray;
+
+
+	/**
+	 * PHP magic method.
+	 * This method allows calling static method defined in [[modelClass]] via this query object.
+	 * It is mainly implemented for supporting the feature of scope.
+	 * @param string $name the method name to be called
+	 * @param array $params the parameters passed to the method
+	 * @return mixed the method return result
+	 */
+	public function __call($name, $params)
+	{
+		if (method_exists($this->modelClass, $name)) {
+			array_unshift($params, $this);
+			call_user_func_array([$this->modelClass, $name], $params);
+			return $this;
+		} else {
+			return parent::__call($name, $params);
+		}
+	}
+
+	/**
+	 * Sets the [[asArray]] property.
+	 * @param boolean $value whether to return the query results in terms of arrays instead of Active Records.
+	 * @return static the query object itself
+	 */
+	public function asArray($value = true)
+	{
+		$this->asArray = $value;
+		return $this;
+	}
+
+	/**
+	 * Specifies the relations with which this query should be performed.
+	 *
+	 * The parameters to this method can be either one or multiple strings, or a single array
+	 * of relation names and the optional callbacks to customize the relations.
+	 *
+	 * The followings are some usage examples:
+	 *
+	 * ~~~
+	 * // find customers together with their orders and country
+	 * Customer::find()->with('orders', 'country')->all();
+	 * // find customers together with their country and orders of status 1
+	 * Customer::find()->with([
+	 *     'orders' => function($query) {
+	 *         $query->andWhere('status = 1');
+	 *     },
+	 *     'country',
+	 * ])->all();
+	 * ~~~
+	 *
+	 * @return static the query object itself
+	 */
+	public function with()
+	{
+		$this->with = func_get_args();
+		if (isset($this->with[0]) && is_array($this->with[0])) {
+			// the parameter is given as an array
+			$this->with = $this->with[0];
+		}
+		return $this;
+	}
+
+	/**
+	 * Sets the [[indexBy]] property.
+	 * @param string|callable $column the name of the column by which the query results should be indexed by.
+	 * This can also be a callable (e.g. anonymous function) that returns the index value based on the given
+	 * row or model data. The signature of the callable should be:
+	 *
+	 * ~~~
+	 * // $model is an AR instance when `asArray` is false,
+	 * // or an array of column values when `asArray` is true.
+	 * function ($model)
+	 * {
+	 *     // return the index value corresponding to $model
+	 * }
+	 * ~~~
+	 *
+	 * @return static the query object itself
+	 */
+	public function indexBy($column)
+	{
+		return parent::indexBy($column);
+	}
+
+	private function createModels($rows)
+	{
+		$models = [];
+		if ($this->asArray) {
+			if ($this->indexBy === null) {
+				return $rows;
+			}
+			foreach ($rows as $row) {
+				if (is_string($this->indexBy)) {
+					$key = $row[$this->indexBy];
+				} else {
+					$key = call_user_func($this->indexBy, $row);
+				}
+				$models[$key] = $row;
+			}
+		} else {
+			/** @var $class ActiveRecord */
+			$class = $this->modelClass;
+			if ($this->indexBy === null) {
+				foreach ($rows as $row) {
+					$models[] = $class::create($row);
+				}
+			} else {
+				foreach ($rows as $row) {
+					$model = $class::create($row);
+					if (is_string($this->indexBy)) {
+						$key = $model->{$this->indexBy};
+					} else {
+						$key = call_user_func($this->indexBy, $model);
+					}
+					$models[$key] = $model;
+				}
+			}
+		}
+		return $models;
+	}
+
+	private function populateRelations(&$models, $with)
+	{
+		$primaryModel = new $this->modelClass;
+		$relations = $this->normalizeRelations($primaryModel, $with);
+		foreach ($relations as $name => $relation) {
+			if ($relation->asArray === null) {
+				// inherit asArray from primary query
+				$relation->asArray = $this->asArray;
+			}
+			$relation->findWith($name, $models);
+		}
+	}
+
+	/**
+	 * @param ActiveRecord $model
+	 * @param array $with
+	 * @return ActiveRelation[]
+	 */
+	private function normalizeRelations($model, $with)
+	{
+		$relations = [];
+		foreach ($with as $name => $callback) {
+			if (is_integer($name)) {
+				$name = $callback;
+				$callback = null;
+			}
+			if (($pos = strpos($name, '.')) !== false) {
+				// with sub-relations
+				$childName = substr($name, $pos + 1);
+				$name = substr($name, 0, $pos);
+			} else {
+				$childName = null;
+			}
+
+			$t = strtolower($name);
+			if (!isset($relations[$t])) {
+				$relation = $model->getRelation($name);
+				$relation->primaryModel = null;
+				$relations[$t] = $relation;
+			} else {
+				$relation = $relations[$t];
+			}
+
+			if (isset($childName)) {
+				$relation->with[$childName] = $callback;
+			} elseif ($callback !== null) {
+				call_user_func($callback, $relation);
+			}
+		}
+		return $relations;
+	}
+}
